pyproject.toml
[tool.poetry.dependencies]
python = "^3.9"
caikit-nlp = { git = "https://github.com/opendatahub-io/caikit-nlp", rev = "0.2.1" }
protobuf  ="4.24.4"
alchemy-config="1.1.2"
alchemy-logging="1.2.1"
grpcio = "1.59.0"
grpcio-reflection = "1.59.0"
#build time dependencies
grpcio-tools = "1.59.0"

[tool.poetry]
name = "caikit_nlp_client"
version = "0.0.1"
description = "Caikit Client"
license =  "Apache-2.0"
readme = "README.md"
classifiers=[
    "License :: OSI Approved :: Apache Software License"
]
authors = ["Vaibhav Jain <vajain@redhat.com>", "Laurentiu Bradin <laurentiu.bradin@ibm.com>"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

